improve delete support on push
* server: implement delete-refs and announce it.
* remote: check if server announced delete-refs before trying
  to delete and fail fast if it does not.

Note that the client does not need no send 'delete-refs' back
to the server to be able to delete references:

```
delete-refs
-----------

If the server sends back the 'delete-refs' capability, it means that
it is capable of accepting a zero-id value as the target
value of a reference update.  It is not sent back by the client, it
simply informs the client that it can be sent zero-id values
to delete references.
```

So our server implementation does not check if the client sent
delete-refs back, it just accepts deletes if it receives them.