Enforce the use of cache in packfile decoder
Decoder object can make use of an object cache to speed up processing.
Previously the only way to specify it was changing manually the struct
generated by NewDecodeForFile. This lead to some instances to be created
without it and penalized performance.

Now the cache should be explicitly passed to the constructor function.
NewDecoder now creates objects with a cache using the default size.

A new helper function was added to create cache objects with the default
size as this becomes a common task now:

    cache.NewObjectLRUDefault()

Signed-off-by: Javi Fontan <jfontan@gmail.com>