plumbing: format/packfile, fix panic retrieving object hash.
In some cases the original data is not saved before it is cleaned
and forces a panic when it's needed.

The change adds ObjectToPack.CleanOriginal to be used to clean original
object instead of:

    object.Original = nil

Now when the Original data is freed because it's no longer in the pack
window a SetOriginal call is done to make sure that Size, Hash and Size
data is not lost.

Signed-off-by: Javi Fontan <jfontan@gmail.com>