NoErrAlreadyUpToDate: Rename to ErrAlreadyUpToDate
I've been unable to find an obvious reason
why NoErrAlreadyUpToDate is named that
instead of following conventions from the rest of the code
and being named ErrAlreadyUpToDate.

This change renames NoErrAlreadyUpToDate to ErrAlreadyUpToDate.

In the interest of backwards compatibility,
we keep the NoErrAlreadyUpToDate variable around,
but mark it as deprecated in favor of ErrAlreadyUpToDate.