Add support for custom git %prefix variables
Git has a quite complex lookup mechanism to find the relevant system
files that is partially hard-coded at compilation time.

Trying to detect it from this library and support all the cases is going to be hard.
Though there is an easy action to take, to allow the library users to
specify the prefix with which git has been compiled.

This commit aims at allowing this by implementing the git documentation
https://git-scm.com/docs/git-config#Documentation/git-config.txt-pathname
and the (undocumented) special case when the prefix is set to /usr
https://github.com/git/git/blob/e773545c7fe7eca21b134847f4fc2cbc9547fa14/Makefile#L1274