Add retry logic for git clone
Git clone was failing for public repository if a wrong authentication
was provided which ideally should not happen because public
repositories does not require authentication.

This change port the same behavior of git CLI that is always try cloning
the repository without authentication, if it fails with 401
retry with authentication.

Signed-off-by: shahulsonhal <shahulsonhal@gmail.com>