git: worktree: check for empty parent dirs during Reset (Fixes #670)
When we delete dir1/dir2/file1, we currently check if dir2 becomes
empty with the deletion of file1, and if so, we delete dir2.  If
dir1 becomes empty with the deletion of dir2, we don't notice that,
and dir1 is left behind.

This commit adds a loop to check each parent directory in the file
path for emptiness, removing empty directories along the way until
a non-empty directory is found (or an error occurs).