git: remote, flip branch check to not-tag check
In remote.updateLocalReferenceStorage, this commit updates a check to
see if the reference is for a branch (in refs/heads) to checking the
reference is not a tag. This change ensures that the subsequent
fast-forward only handling clauses apply to references that are not
standard branches stored in refs/heads.

Signed-off-by: Aditya Sirish <aditya@saky.in>