storage: filesystem, handle submodule names with encoding
Whilst trying to get to the bottom of unexpected mutative behaviour
from calling `submodule.Status()` it was determined that submodule names
containing characters that Git URL encodes (such as forward slashes on
Linux) were not being handled correctly, resulting in
`submodule.Repository()` calling Init() with an incorrect submodule name.

Further, Go's URL encoding opts for uppercase hexadecimal characters,
whereas Git appears to go for lowercase, so some hacky handling of this
is required.

We haven't been able to locate canonical encoding/decoding behaviour in
the Git source, so it is currently unknown how comprehensive this is. We
have opted for the obvious suspects, being forward and backslashes.

Co-authored-by: Rex Pan <rex@google.com>