config: implement v6 LoadConfig alongside v5 LoadConfig
I have discovered a bug in the go-git v5 implementation of loading
git config in that it does not behave as expected according to the
git config documentation, specifically in how it loads local config
only from the in-memory config storer by default, in contrast to
the documented behavior that by default git should load and respect
a merged configuration by layering values from several convention-
based file locations, including "global" config typically located
at $HOME/.gitconfig and "system" config typically located at
/etc/gitconfig

There are likely additional bugs hiding in the v5 implementation
of merging config values as the src and dest parameters to the
mergo.Merge(...) function appear to be opposite from expected;
based on the documentation of the mergo.Merge(...) command I
expect src to get merged into dest but the usage in
repository.ConfigScoped(..) looks like it is trying to merge
dest into src.

In discussion with project maintainers (see #395) it was determined
that while the go-git v5 implementation diverges from what clients
would expect, fixing it constitutes a breaking change for clients
who have coded to the existing v5 config behavior.

This commit introduces one possible method to support v5 and v6
config implementations side-by-side in a non-breaking way, allowing
clients to opt-in to v6 behavior ahead of the go-git v6 release.

Applying this commit adds new options to the config.Scope enum
including
- V6DefaultScope
- V6SystemScope
- V6GlobalScope
- V6LocalScope

Config path resolution is implemented according to the description
section of the current git 2.43 documentation for the config
subcommand:

    When reading, the values are read from the system, global
    and repository local configuration files by default, and
    options --system, --global, --local, --worktree and
    --file <filename> can be used to tell the command to read
    from only that location (see FILES).

    - https://git-scm.com/docs/git-config/2.43.0#_description

And the related FILES section

    By default, git config will read configuration options from
    multiple files:

    $(prefix)/etc/gitconfig
    System-wide configuration file.

    $XDG_CONFIG_HOME/git/config
    ~/.gitconfig
    User-specific configuration files. When the XDG_CONFIG_HOME
    environment variable is not set or empty, $HOME/.config/ is
    used as $XDG_CONFIG_HOME.

    These are also called "global" configuration files. If both
    files exist, both files are read in the order given above.

    $GIT_DIR/config
    Repository specific configuration file.

    $GIT_DIR/config.worktree
    This is optional and is only searched when
    extensions.worktreeConfig is present in $GIT_DIR/config.

    You may also provide additional configuration parameters when
    running any git command by using the -c option. See git[1]
    for details.

    Options will be read from all of these files that are available.
    If the global or the system-wide configuration files are missing
    or unreadable they will be ignored. If the repository configuration
    file is missing or unreadable, git config will exit with a non-zero
    error code. An error message is produced if the file is unreadable,
    but not if it is missing.

    The files are read in the order given above, with last value
    found taking precedence over values read earlier. When multiple
    values are taken then all values of a key from all files will
    be used.

    - https://git-scm.com/docs/git-config/2.43.0#FILES

This commit uses the existing mergo library to load either each
of these files as a sequence based on which V6 scope is requested
and merge them all into a single merged config.

NOTE: This commit does not handle the --worktree nor --file <filename>
      flags as both options require a more extensive refactoring or
      reimplementation of the config system to access repository
      configuration while resolving config values in the --worktreee case
      and accept and pass the <filename> parameter to the config module
      in the --file case.

      Support for these options can be reconsidered in go-git v6
      planning.