plumbing: object, don't traverse past shallow commits in bfs commit walk
When trying to determine the merge base between two commits in a shallow
clone, bfsCommitIterator tries to traverse the entire history until it
reaches the shallow point and tries to look up the parent commits, which
do not exist in the shallow clone.

This causes `commit.MergeBase` to return a 'object not found' error.

Handle this by obtaining a list of shallow commits, if available.

This fixes an issue with commit.MergeBase, that git proper handles
without problems.