plumbing: transport, determine protocol version
This adds two functions to the transport package to determine the
protocol version for the client and server. The DiscoverVersion function
reads the first pktline from the reader to determine the protocol
version. This is usually used by the client-side.
The ProtocolVersion function tries to find the version parameter in the
protocol string. This expects the protocol string from the GIT_PROTOCOL
environment variable. This is used by the server-side.