Add link to comment about the "legal issues"
Why? Because some people see "legal issues" and think "uh oh... a dark cloud is looming over this project now." Providing a link to more context to dispel the reflexive negative reaction to the words "legal issues" seems like a good idea. The GitHub comment was the best that I could find. If there's a better summary somewhere then update this link.