plumbing: fix, prevent double decoding of commands and flush
This refactors the decoding of commands and capabilities in the
update-requests decoder to avoid double decoding the commands
and eating the next pktlines in the reader.

The issue was that the decoder doesn't stop once it reaches a flush
packet and continues to read the next line. This is incorrect as
an update-requests always expects a flush packet after the commands
have been sent.