plumbing: transport, fix advertise protocol version
This commit fixes the protocol version advertisement for the ReceivePack
and UploadPack services. The version should be advertised during the
advertisement phase if and only if we're advertising references during
the ReceivePack or UploadPack routines.

This will also change the behavior of the AdvertiseReferences function to
only accept UploadPackService and ReceivePackService as valid services.

It will also error out if the protocol version is not supported.