plumbing: transport, packp, properly implement push-options
The Git protocol defines push-options as simple strings. Whether
they are key-value pairs or simple strings it is up to the
client. This reflects that in the codebase.

It also refactors the push-options logic out of the UpdateRequests
struct type and into its own type. This allows us to encode and decode
push-options separately from the update requests. Previously,
push-options were only encoded and the decoding logic was missing. This
is probably because before we were relying on the git binary to do that
for us when using the file transport i.e. server-side.