plumbing: transport/ssh allow different AuthMethod basd on user/host
When using the plain git with SSH, git reads the ssh config and is able to consider
all the relevant SSH options depending on the hostname.

With go-git, it is not possible to use different configurations like IdentityFiles
based on the hostname or other configurations as the information is not provided to the
ssh auth config.

This commit provides such an information while keeping the backward compatibility with
older auth methods which do not.