git: allow pushing to an anonymous remote
Git allows you to push to a remote that isn't defined in your
gitconfig:

    git push https://example.com/foo.git refs/heads/bar:refs/heads/bar

The effect of this command is to push to the remote repository at the
specified URL, without updating any remote-tracking references
corresponding to the reference that was pushed.

It would seem that this should be possible with `go-git`, by writing
something like

    r, _ := git.PlainOpen(".")
    _ = r.Push(&git.PushOptions{
        RemoteURL: "https://example.com/foo.git",
        RefSpecs: []config.RefSpec{"refs/heads/bar:refs/heads/bar"},
    })

But `go-git` wasn't handling this correctly:

* `PushOptions.Validate()` saw the empty `o.RemoteName` and changed it
  to `DefaultRemoteName` ("origin").

* If no "origin" remote was defined in the gitconfig,
  `Repository.Push()` failed entirely with `ErrRemoteNotFound`.

* If: an "origin" _was_ defined in the gitconfig, then after the push,
  remote-tracking references were created for the "origin"
  remote (e.g., `refs/remotes/origin/foo`), even though the push was
  probably to a remote that is not related to "origin".

Alternatively, if the `PushOptions` were created using `RemoteName:
"anonymous"`, the push still failed with `ErrRemoteNotFound` because a
remote called "anonymous" was sought in the gitconfig, and that
remote:(presumably!) doesn't exist.

This commit fixes all three problems:

* Change the behavior of `PushOptions.Validate()`: if there is a
  `RemoteURL` but no `RemoteName`, then set the `RemoteName` to
  `config.AnonymousRemoteName` rather than setting it to
  `DefaultRemoteName`. (If neither a `RemoteName` nor a `RemoteURL` is
  set, then set `RemoteName` to `DefaultRemoteName` as before.)

* Change `Remote.updateRemoteReferenceStorage()` to skip updating any
  remote-tracking references if the remote is anonymous.