Fix TestCheckoutRelativePathSubmoduleInitialized commit options
Add missing Author field to CommitOptions in test to align with other
tests in the codebase. This test was missed in commit 473d9879 where
other similar tests were updated to use proper commit signatures.

The test was failing due to missing git author configuration, which
is required for commit operations in go-git v6.

Signed-off-by: cappyzawa <cappyzawa@gmail.com>