utils: sync, Fix zlib reader test
Putting empty byte array to the pool has potential to interrupt other tests
as it isn't intended size(32kb). Therefore use byte array from the pool instead
of creating one.

Also all the tests in this package is flaky. sync.Pool doesn't ensure that the
object we put in will still exist until we get it back. Document the behavior
for now, so we can decide whether to remove the tests, or consider other approaches.